chore: refactor on ClientConfig.ts file
