+++ src/sdk/Clients/Vulnerability/VulnerabilityHttpClient.ts
@@ -1,8 +1,9 @@
-import { ClientConfig, HttpClient } from "../http/HttpClient";
+import { HttpClient } from "../http/HttpClient";
 import { IVulnerabilityClient } from "./IVulnerabilityClient";
 import { logger } from "../../Logger";
 import { Component } from "../../types/common/types";
 import { ComponentsVulnerabilitiesResponse, ComponentVulnerabilityResponse } from "../../types/vulnerability/types";
+import { ClientConfig } from "../interfaces/ClientConfig";
 
 export class VulnerabilityHttpClient extends HttpClient implements IVulnerabilityClient{
