+++ src/sdk/Vulnerability/VulnerabilityScanner.ts
@@ -8,7 +8,7 @@ import {
   ComponentVulnerabilityResponse
 } from "../types/vulnerability/types";
 import { Component } from "../types/common/types";
-import { ClientConfig } from "../Clients/http/HttpClient";
+import { ClientConfig } from "../Clients/interfaces/ClientConfig";
 
 export class VulnerabilityScanner {
   private config: VulnerabilityCfg;
